Moved session/index.js -> session/middleware.js
